Introduced stream read & write traits.
