+++ src/Stream/Writer.php
@@ -24,15 +24,11 @@ use Concurrent\Watcher;
 
 class Writer implements WritableStream
 {
+    use WriterTrait;
+    
     protected $resource;
     
     protected $watcher;
-    
-    protected $bufferSize;
-    
-    protected $buffer = '';
-    
-    protected $writing = false;
     
     public function __construct($resource, ?Watcher $watcher = null)
     {
@@ -45,58 +41,10 @@ class Writer implements WritableStream
     
     public function close(?\Throwable $e = null): void
     {
-        if ($this->resource !== null) {
-            $meta = @\stream_get_meta_data($this->resource);
-            
-            if ($meta && \strpos($meta['mode'], '+') !== false) {
-                @\stream_socket_shutdown($this->resource, \STREAM_SHUT_WR);
-            } else {
-                @\fclose($this->resource);
-            }
+        if (\is_resource($this->resource)) {
+            $this->closeWriter($e);
             
             $this->resource = null;
-            
-            $this->watcher->close($e);
-        }
-    }
-
-    public function write(string $data): void
-    {
-        $retried = false;
-        
-        while ($this->writing) {
-            $this->watcher->awaitWritable();
-        }
-        
-        while ($data !== '') {
-            if (!\is_resource($this->resource)) {
-                throw new StreamClosedException('Cannot write to closed stream');
-            }
-            
-            if (false === ($len = @\fwrite($this->resource, $data, 0xFFFF))) {
-                throw new StreamException(\sprintf('Could not write to stream: %s', \error_get_last()['message'] ?? ''));
-            }
-            
-            if ($len > 0) {
-                $data = \substr($data, $len);
-                $retried = false;
-            } elseif (@\feof($this->resource)) {
-                throw new StreamClosedException('Cannot write to closed stream');
-            } else {
-                if ($retried) {
-                    throw new StreamClosedException('Could not write bytes after retry, assuming broken pipe');
-                }
-                
-                $this->writing = true;
-                
-                try {
-                    $this->watcher->awaitWritable();
-                } finally {
-                    $this->writing = false;
-                }
-                
-                $retried = true;
-            }
         }
     }
 }
